- reference resolving visitor is inner type of more general type resolver - type visitor promoted to public api - better code style for assertion
