+++ src/main/java/graphql/schema/GraphQLTypeResolvingVisitor.java
@@ -1,6 +1,6 @@
 package graphql.schema;
 
-import graphql.Assert;
+import static graphql.Assert.assertNotNull;
 import graphql.Internal;
 import graphql.util.TraversalControl;
 import graphql.util.TraverserContext;
@@ -38,8 +38,47 @@ public class GraphQLTypeResolvingVisitor extends GraphQLTypeVisitorStub {
     public TraversalControl visitGraphQLTypeReference(GraphQLTypeReference node, TraverserContext<GraphQLType> context) {
 
         final GraphQLType resolvedType = typeMap.get(node.getName());
-        Assert.assertTrue(resolvedType != null, "type %s not found in schema", node.getName());
-        context.getParentContext().thisNode().accept(context, new GraphQLTypeRefResolvingVisitor(resolvedType));
+        assertNotNull(resolvedType != null, "type %s not found in schema", node.getName());
+        context.getParentContext().thisNode().accept(context, new TypeRefResolvingVisitor(resolvedType));
         return super.visitGraphQLTypeReference(node, context);
+    }
+
+
+    private class TypeRefResolvingVisitor extends GraphQLTypeVisitorStub {
+        protected final GraphQLType resolvedType;
+
+        TypeRefResolvingVisitor(GraphQLType resolvedType) {
+            this.resolvedType = resolvedType;
+        }
+
+        @Override
+        public TraversalControl visitGraphQLFieldDefinition(GraphQLFieldDefinition node, TraverserContext<GraphQLType> context) {
+            node.replaceType((GraphQLOutputType) resolvedType);
+            return super.visitGraphQLFieldDefinition(node, context);
+        }
+
+        @Override
+        public TraversalControl visitGraphQLArgument(GraphQLArgument node, TraverserContext<GraphQLType> context) {
+            node.replaceType((GraphQLInputType) resolvedType);
+            return super.visitGraphQLArgument(node, context);
+        }
+
+        @Override
+        public TraversalControl visitGraphQLInputObjectField(GraphQLInputObjectField node, TraverserContext<GraphQLType> context) {
+            node.replaceType((GraphQLInputType) resolvedType);
+            return super.visitGraphQLInputObjectField(node, context);
+        }
+
+        @Override
+        public TraversalControl visitGraphQLList(GraphQLList node, TraverserContext<GraphQLType> context) {
+            node.replaceType(resolvedType);
+            return super.visitGraphQLList(node, context);
+        }
+
+        @Override
+        public TraversalControl visitGraphQLNonNull(GraphQLNonNull node, TraverserContext<GraphQLType> context) {
+            node.replaceType(resolvedType);
+            return super.visitGraphQLNonNull(node, context);
+        }
     }
 }
